[WIP] Multiple fixes (but still memory leaking): see commit description
The segfault that was caused by not reassigning the packet reference to nullptr when they were killed in PacketBatch:kill, resulting in improperly freed memory and eventually a segfault. This explains why the time to a segfault was dependent on the memorypool's initial size.

When inspecting the program with valgrind, there is still a massive memory leak that comes from the memorypool. I still have to investigate why it occurs, though my current hypothesis is that since the pool lives during the whole program's life time, it may be seen as "definitely lost" by valgrind since it's destructor might not be called.

Changelog :

- fix segfault on PacketBatchVector::kill, by popping the packet from the internal packet pointer array when the vector is killed. The batch is also released from the memory.
- change memory_pool to be shared among all
- PacketBatchVector instances
- reimplement EXECUTE_FOR_EACH_PACKET_UNTIL_DO_VEC, EXECUTE_FOR_EACH_PACKET_UNTIL_DROP_VEC, MAKE_BATCH_VEC for vector
- Suppress compilation for unused variable in unused functions (might change)
- Add a check if the batch is full in append_packet
- make_tail now returns this, as this is only kept for legacy reasons, same for start_head, which calls make_from_packet directly
- correct logic in cut and split
- add a function to pop a packet from the internal array (it does not shift the remaining packets to the left, see function specifications).
- rewrite clone_batch for vector
- rewrite BATCH_CREATE_INIT, BATCH_CREATE_APPEND, BATCH_CREATE_FINISH for vector

Signed-off-by: Emile Villette <emile.villette@student.uclouvain.be>